feat: add new model (ProductManufacturingInfo) to get both product and drawing from product-service by only one request

+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/models/FactoryOrder.java
@@ -13,10 +13,10 @@ import org.springframework.data.redis.core.RedisHash;
 public class FactoryOrder {
 
   @Id
-  private Long orderId;
+  private int orderId;
 
-  private Long productId;
+  private int productId;
 
-  private Integer productQuantity;
+  private int productQuantity;
 
 }
