+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/models/OrderCreatedEvent.java
@@ -2,7 +2,7 @@ package com.veteroch4k.factory_service.models;
 
 
 public record OrderCreatedEvent(
-    Long orderId,
-    Long productId,
-    Integer productQuantity
+    int orderId,
+    int productId,
+    int productQuantity
 ) {}
